Move search routines into a dedicated source file
