+++ imap/imap.c
@@ -334,157 +334,6 @@ static int sync_helper(struct Mailbox *m, AclFlags right, int flag, const char *
   return count;
 }
 
-// fwd-declaration, check_pattern and check_pattern_list are mutually recursive.
-static int check_pattern_list(const struct PatternList *patterns);
-
-/**
- * check_pattern - Check whether a pattern can be executed server-side
- * @param pat Pattern to check
- * @retval True if pattern can be done server-side
- */
-static bool check_pattern(const struct Pattern *pat)
-{
-  switch (pat->op)
-  {
-    case MUTT_PAT_BODY:
-    case MUTT_PAT_HEADER:
-    case MUTT_PAT_WHOLE_MSG:
-      if (pat->string_match)
-        return true;
-      break;
-    case MUTT_PAT_SERVERSEARCH:
-      return true;
-      break;
-    default:
-      if (pat->child && check_pattern_list(pat->child))
-        return true;
-      break;
-  }
-  return false;
-}
-
-/**
- * check_pattern_list - Check how many patterns in a list can be executed server-side
- * @param search  List of pattern to match
- * @retval num Number of patterns search that can be done server-side
- */
-static int check_pattern_list(const struct PatternList *patterns)
-{
-  int positives = 0;
-
-  const struct Pattern *pat = NULL;
-  SLIST_FOREACH(pat, patterns, entries)
-  {
-    positives += check_pattern(pat);
-  }
-
-  return positives;
-}
-
-/**
- * compile_search - Convert NeoMutt pattern to IMAP search
- * @param adata Imap Account data
- * @param pat Pattern to convert
- * @param buf Buffer for result
- * @retval True on success
- * @retval False on failure
- *
- * Convert neomutt Pattern to IMAP SEARCH command containing only elements
- * that require full-text search (neomutt already has what it needs for most
- * match types, and does a better job (eg server doesn't support regexes).
- */
-static bool compile_search(const struct ImapAccountData *adata,
-                           const struct Pattern *pat, struct Buffer *buf)
-{
-  if (!check_pattern(pat))
-    return true;
-
-  if (pat->pat_not)
-    mutt_buffer_addstr(buf, "NOT ");
-
-  if (pat->child)
-  {
-    int clauses = check_pattern_list(pat->child);
-    if (clauses > 0)
-    {
-      mutt_buffer_addch(buf, '(');
-
-      struct Pattern *c = SLIST_FIRST(pat->child);
-      while (clauses)
-      {
-        if (check_pattern(c))
-        {
-          if ((pat->op == MUTT_PAT_OR) && (clauses > 1))
-            mutt_buffer_addstr(buf, "OR ");
-          clauses--;
-
-          if (!compile_search(adata, c, buf))
-            return false;
-
-          if (clauses)
-            mutt_buffer_addch(buf, ' ');
-        }
-        c = SLIST_NEXT(c, entries);
-      }
-
-      mutt_buffer_addch(buf, ')');
-    }
-  }
-  else
-  {
-    char term[256];
-    char *delim = NULL;
-
-    switch (pat->op)
-    {
-      case MUTT_PAT_HEADER:
-        mutt_buffer_addstr(buf, "HEADER ");
-
-        /* extract header name */
-        delim = strchr(pat->p.str, ':');
-        if (!delim)
-        {
-          mutt_error(_("Header search without header name: %s"), pat->p.str);
-          return false;
-        }
-        *delim = '\0';
-        imap_quote_string(term, sizeof(term), pat->p.str, false);
-        mutt_buffer_addstr(buf, term);
-        mutt_buffer_addch(buf, ' ');
-
-        /* and field */
-        *delim = ':';
-        delim++;
-        SKIPWS(delim);
-        imap_quote_string(term, sizeof(term), delim, false);
-        mutt_buffer_addstr(buf, term);
-        break;
-      case MUTT_PAT_BODY:
-        mutt_buffer_addstr(buf, "BODY ");
-        imap_quote_string(term, sizeof(term), pat->p.str, false);
-        mutt_buffer_addstr(buf, term);
-        break;
-      case MUTT_PAT_WHOLE_MSG:
-        mutt_buffer_addstr(buf, "TEXT ");
-        imap_quote_string(term, sizeof(term), pat->p.str, false);
-        mutt_buffer_addstr(buf, term);
-        break;
-      case MUTT_PAT_SERVERSEARCH:
-        if (!(adata->capabilities & IMAP_CAP_X_GM_EXT_1))
-        {
-          mutt_error(_("Server-side custom search not supported: %s"), pat->p.str);
-          return false;
-        }
-        mutt_buffer_addstr(buf, "X-GM-RAW ");
-        imap_quote_string(term, sizeof(term), pat->p.str, false);
-        mutt_buffer_addstr(buf, term);
-        break;
-    }
-  }
-
-  return true;
-}
-
 /**
  * longest_common_prefix - Find longest prefix common to two strings
  * @param dest  Destination buffer
@@ -1351,38 +1200,6 @@ int imap_mailbox_status(struct Mailbox *m, bool queue)
   if (!adata || !mdata)
     return -1;
   return imap_status(adata, mdata, queue);
-}
-
-/**
- * imap_search - Find messages in mailbox matching a pattern
- * @param m   Mailbox
- * @param pat Pattern to match
- * @retval true  Success
- * @retval false Failure
- */
-bool imap_search(struct Mailbox *m, const struct PatternList *pat)
-{
-  for (int i = 0; i < m->msg_count; i++)
-  {
-    struct Email *e = m->emails[i];
-    if (!e)
-      break;
-    e->matched = false;
-  }
-
-  if (check_pattern_list(pat) == 0)
-    return true;
-
-  struct Buffer buf;
-  mutt_buffer_init(&buf);
-  mutt_buffer_addstr(&buf, "UID SEARCH ");
-
-  struct ImapAccountData *adata = imap_adata_get(m);
-  const bool ok = compile_search(adata, SLIST_FIRST(pat), &buf) &&
-                  (imap_exec(adata, buf.data, IMAP_CMD_NO_FLAGS) == IMAP_EXEC_SUCCESS);
-
-  FREE(&buf.data);
-  return ok;
 }
 
 /**
