+++ imap/lib.h
@@ -41,6 +41,7 @@
  * | imap/browse.c     | @subpage imap_browse     |
  * | imap/command.c    | @subpage imap_command    |
  * | imap/message.c    | @subpage imap_message    |
+ * | imap/search.c     | @subpage imap_search     |
  * | imap/utf7.c       | @subpage imap_utf7       |
  * | imap/util.c       | @subpage imap_util       |
  */
@@ -93,7 +94,6 @@ int imap_delete_mailbox(struct Mailbox *m, char *path);
 int imap_sync_mailbox(struct Mailbox *m, bool expunge, bool close);
 int imap_path_status(const char *path, bool queue);
 int imap_mailbox_status(struct Mailbox *m, bool queue);
-bool imap_search(struct Mailbox *m, const struct PatternList *pat);
 int imap_subscribe(char *path, bool subscribe);
 int imap_complete(char *buf, size_t buflen, const char *path);
 int imap_fast_trash(struct Mailbox *m, char *dest);
@@ -124,5 +124,8 @@ void imap_keepalive(void);
 
 void imap_get_parent_path(const char *path, char *buf, size_t buflen);
 void imap_clean_path(char *path, size_t plen);
+
+/* search.c */
+bool imap_search(struct Mailbox *m, const struct PatternList *pat);
 
 #endif /* MUTT_IMAP_LIB_H */
